+++ src/universal/modules/teamDashboard/containers/TeamSettings/TeamSettingsContainer.js
@@ -1,4 +1,4 @@
-import React, {Component, PropTypes} from 'react';
+import React, {PropTypes} from 'react';
 import {connect} from 'react-redux';
 import TeamSettings from 'universal/modules/teamDashboard/components/TeamSettings/TeamSettings';
 import {cashay} from 'cashay';
@@ -12,39 +12,88 @@ query {
   },
   teamMembers(teamId: $teamId) @live {
     id
+    email
     isLead
     picture
     preferredName
+  },
+  invitations(teamId: $teamId) @live {
+    id
+    email
+    tokenExpiration
+    updatedAt
   }
 }`;
 
 const mapStateToProps = (state, props) => {
   const {teamId} = props.params;
-  const {team, teamMembers} = cashay.query(teamSettingsQuery, {
+  const {invitations, team, teamMembers} = cashay.query(teamSettingsQuery, {
     op: 'teamSettingsContainer',
     key: teamId,
+    sort: {
+      teamMembers: (a, b) => a.preferredName > b.preferredName ? 1 : -1,
+      invitations: (a, b) => a.createdAt > b.createdAt ? 1 : -1
+    },
     variables: {teamId}
   }).data;
   return {
+    invitations,
     team,
-    teamMembers
+    teamMembers,
+    leaveTeamModal: state.teamSettings.leaveTeamModal,
+    promoteTeamMemberModal: state.teamSettings.promoteTeamMemberModal,
+    removeTeamMemberModal: state.teamSettings.removeTeamMemberModal,
+    modalTeamMemberId: state.teamSettings.teamMemberId,
+    modalPreferredName: state.teamSettings.preferredName,
+    myTeamMemberId: `${state.auth.obj.sub}::${teamId}`
   };
 };
 
-@connect(mapStateToProps)
-// eslint-disable-next-line react/prefer-stateless-function
-export default class TeamSettingsContainer extends Component {
-  static propTypes = {
-    params: PropTypes.object.isRequired,
-    team: PropTypes.object.isRequired,
-    teamMembers: PropTypes.array.isRequired
-  };
-
-  render() {
-    const {team, teamMembers} = this.props;
-    if (teamMembers.length === 0) {
-      return <LoadingView/>;
-    }
-    return <TeamSettings team={team} teamMembers={teamMembers}/>;
+const TeamSettingsContainer = (props) => {
+  const {
+    dispatch,
+    invitations,
+    leaveTeamModal,
+    modalPreferredName,
+    modalTeamMemberId,
+    myTeamMemberId,
+    promoteTeamMemberModal,
+    removeTeamMemberModal,
+    team,
+    teamMembers
+  } = props;
+  const myTeamMember = teamMembers.find((member) => member.id === myTeamMemberId);
+  if (!myTeamMember) {
+    return <LoadingView/>;
   }
-}
+  return (
+    <TeamSettings
+      dispatch={dispatch}
+      invitations={invitations}
+      leaveTeamModal={leaveTeamModal}
+      myTeamMember={myTeamMember}
+      modalTeamMemberId={modalTeamMemberId}
+      modalPreferredName={modalPreferredName}
+      promoteTeamMemberModal={promoteTeamMemberModal}
+      removeTeamMemberModal={removeTeamMemberModal}
+      team={team}
+      teamMembers={teamMembers}
+    />
+  );
+};
+
+TeamSettingsContainer.propTypes = {
+  dispatch: PropTypes.func.isRequired,
+  invitations: PropTypes.array.isRequired,
+  leaveTeamModal: PropTypes.bool.isRequired,
+  modalPreferredName: PropTypes.string,
+  modalTeamMemberId: PropTypes.string,
+  myTeamMemberId: PropTypes.string.isRequired,
+  params: PropTypes.object.isRequired,
+  promoteTeamMemberModal: PropTypes.bool.isRequired,
+  removeTeamMemberModal: PropTypes.bool.isRequired,
+  team: PropTypes.object.isRequired,
+  teamMembers: PropTypes.array.isRequired
+};
+
+export default connect(mapStateToProps)(TeamSettingsContainer);