+++ src/universal/routes/teamDashMain.js
@@ -1,4 +1,4 @@
-export default {
+export default (store) => ({
   path: ':teamId',
   getComponent: async(location, cb) => {
     const component = await System.import('universal/modules/teamDashboard/containers/Team/TeamContainer');
@@ -10,20 +10,10 @@ export default {
   },
   getChildRoutes: (childLocation, cbChild) => {
     cbChild(null, [
-      {
-        path: 'archive',
-        getComponent: async (location, cb) => {
-          const component = await System.import('universal/modules/teamDashboard/containers/TeamArchive/TeamArchiveContainer');
-          cb(null, component);
-        }
-      },
-      {
-        path: 'settings',
-        getComponent: async (location, cb) => {
-          const component = await System.import('universal/modules/teamDashboard/containers/TeamSettings/TeamSettingsContainer');
-          cb(null, component);
-        }
-      }
+      /* eslint-disable global-require */
+      require('./teamArchive')(store),
+      require('./teamSettings')(store)
+      /* eslint-enable */
     ]);
   }
-};
+});